taxifyl-staging/taxifyl-staging/requirements.txt
# =============================================================================
# US Tax Form Extractor - Production Dependencies
# =============================================================================

# --- Web Framework ---
fastapi>=0.109.0
uvicorn>=0.27.0
streamlit>=1.30.0

# --- API & Requests ---
requests>=2.31.0
python-multipart>=0.0.6

# --- Database ---
supabase>=2.3.0

# --- AI/ML APIs ---
google-genai>=0.3.0
groq>=0.4.0

# --- PDF Processing ---
PyMuPDF>=1.23.0

# --- Image Processing ---
Pillow>=10.0.0

# --- Data Validation ---
pydantic>=2.5.0

# --- Environment ---
python-dotenv>=1.0.0

# --- Data Processing ---
pandas>=2.0.0
numpy>=1.24.0
